# Conflicts:
#	Erebus/Erebus.vcxproj
#	Erebus/Erebus.vcxproj.filters
#	Erebus/Scripts/gameplay.lua
#	Erebus/Scripts/player.lua
#	Erebus/Scripts/spellList.lua
#	Erebus/source.cpp

+++ Nurn/ChangeSpellsPacket.hpp
@@ -16,7 +16,7 @@ union ChangeSpellsPacket
 	{
 	}
 
-	ChangeSpellsPacket(const uint16_t &spellSlot1, const uint16_t &spellSlot2, const uint16_t &spellSlot3)
+	ChangeSpellsPacket(const uint8_t &spellSlot1, const uint8_t &spellSlot2, const uint8_t &spellSlot3)
 	{
 		data.spellSlot1 = spellSlot1;
 		data.spellSlot2 = spellSlot2;
